Refactor landing page component structure
